::Objects FFI
