+++ 3DRadSpace/Engine3DRadSpace/FFI/Input/Keyboard.h
@@ -12,10 +12,10 @@ typedef struct E3DRSP_Keyboard
 extern "C"
 {
 #endif
-	E3DRSP_INPUT_EXPORT bool E3DRSP_Keyboard_IsKeyDown(E3DRSP_Keyboard* kb, E3DRSP_Key key);
-	E3DRSP_INPUT_EXPORT bool E3DRSP_Keyboard_IsKeyUp(E3DRSP_Keyboard* kb, E3DRSP_Key key);
-	E3DRSP_INPUT_EXPORT bool E3DRSP_Keyboard_IsAnyKeyDown(E3DRSP_Keyboard *kb);
-	E3DRSP_INPUT_EXPORT E3DRSP_KeyState E3DRSP_Keyboard_At(E3DRSP_Keyboard* kb, E3DRSP_Key key);
+	E3DRSP_FFI_EXPORT bool E3DRSP_Keyboard_IsKeyDown(E3DRSP_Keyboard* kb, E3DRSP_Key key);
+	E3DRSP_FFI_EXPORT bool E3DRSP_Keyboard_IsKeyUp(E3DRSP_Keyboard* kb, E3DRSP_Key key);
+	E3DRSP_FFI_EXPORT bool E3DRSP_Keyboard_IsAnyKeyDown(E3DRSP_Keyboard *kb);
+	E3DRSP_FFI_EXPORT E3DRSP_KeyState E3DRSP_Keyboard_At(E3DRSP_Keyboard* kb, E3DRSP_Key key);
 #ifdef __cplusplus
 }
 #endif
