+++ 3DRadSpace/Engine3DRadSpace/FFI/Logging/Message.h
@@ -12,7 +12,7 @@ typedef struct E3DRSP_Message
 extern "C"
 {
 #endif
-	void E3DRSP_LOGGING_EXPORT E3DRSP_SetLastMessage(const E3DRSP_Message* message);
+	void E3DRSP_FFI_EXPORT E3DRSP_SetLastMessage(const E3DRSP_Message* message);
 #ifdef __cplusplus
 }
 #endif
