+++ 3DRadSpace/Engine3DRadSpace/FFI/Logging/Warning.h
@@ -13,7 +13,7 @@ typedef struct E3DRSP_Warning
 extern "C"
 {
 #endif
-	void E3DRSP_LOGGING_EXPORT E3DRSP_SetLastWarning(const E3DRSP_Warning* warning);
+	void E3DRSP_FFI_EXPORT E3DRSP_SetLastWarning(const E3DRSP_Warning* warning);
 #ifdef __cplusplus
 }
 #endif
