+++ 3DRadSpace/Engine3DRadSpace/Logging/CMakeLists.txt
@@ -8,11 +8,9 @@ add_library(3DRadSpace.Logging SHARED
 	"Logging.hpp"
 	"Message.hpp"
 	"Message.cpp"
-	"../FFI/Logging/Message.h"
 	"ServiceBadCastException.hpp"
 	"Warning.hpp"
 	"Warning.cpp"
-	"../FFI/Logging/Warning.h"
 )
 
 if (CMAKE_VERSION VERSION_GREATER 3.12)
