+++ 3DRadSpace/Engine3DRadSpace/Objects/ICamera.cpp
@@ -1,5 +1,4 @@
 #include "ICamera.hpp"
-#include "ICamera.h"
 
 using namespace Engine3DRadSpace;
 using namespace Engine3DRadSpace::Objects;
@@ -17,18 +16,4 @@ ICamera::ICamera(
 	NearPlaneDistance(npd),
 	FarPlaneDistance(fpd)
 {
-}
-
-E3DRSP_Matrix4x4 E3DRSP_ICamera_GetViewMatrix(E3DRSP_ICamera cam)
-{
-	return std::bit_cast<E3DRSP_Matrix4x4>(static_cast<ICamera*>(cam)->GetViewMatrix());
-}
-
-E3DRSP_Matrix4x4 E3DRSP_ICamera_GetProjectionMatrix(E3DRSP_ICamera cam)
-{
-	return std::bit_cast<E3DRSP_Matrix4x4>(static_cast<ICamera*>(cam)->GetProjectionMatrix());
-}
-void E3DRSP_ICamera_Destroy(E3DRSP_ICamera cam)
-{
-	delete static_cast<ICamera*>(cam);
 }
